Switch from methods to yielding for template inheritence
While using methods for accessing content in a template is "clean" in
terms of implementation, Rails-esuqe yielding of sections results in
much nicer templates.

Much of my initial unease with using yield for this feature comes from
Ruby's somewhat lack of flexibility in this circumstance. Since the code
is being instance_eval'd, we need a block in the current binding that
yield will delegate to. However, Ruby doesn't provide a direct way to set
the block of a binding. like Binding#block= or something. Instead, we
have to call out to a separate method which does the eval'ing and pass
in a block to that method; that way, when the compiled ERB is evaluated,
the default block does the yielding of either the main section or a
named section.

Again, this is not the prettiest solution, implementation-wise. However,
the templates read just that much clearer, and the complexity is
contained to this rather isolated ERB engine class, so I'm satisfied
with it.

I wonder if it'd even be possible on the Ruby side to add a method like
Binding#block=, or if that would even be the right solution? My
intuition says that would work, but I could totally be wrong. I also
don't know if it would even be possible to implement on the Ruby side.
Something to look into, I suppose.