Reintroduce template inheritance through a "template chain"
Most templating languages have the concepts of templates which "extend"
other templates, typically indicated through a tag at the top of the
file called "extends" or something of the like. However, in the context
of a static site generator, I've been thinking less and less of pages as
"inheriting from" other templates, and more of them "using" other
templates as their base.

The best word I could come up with to model this relationship was a
"chain", as templates are rendered in order, passing the context along
to each successive one. I think this word might cause confusion with
the pipeline engine, which serves a similar, but unrelated purpose.
Maybe better and more distinct words can be chosen for both, or maybe
they're fine as-is.

Regardless, I'm happy to separate this out from the core Template class.
Having it in there created a weird disconnect between a template just
rendering itself, and a template rendering itself and all its parents.
Separating it out means that a Template is always responsible for
rendering itself, and a Template::Chain is always responsible for
rendering a sequence of Templates.