Add link_to template helper
I'm not entirely satisfied with the Template::Helpers module name;
"helpers" always seems like a bit of a dumping ground for things, like
"support" or "utils". The contents of the module aren't really relevant
outside of generating content to be put in a template, though, so it
at least belongs under the Template class namespace.