Rename build_* factories to create_*
Since computing the output of a resource will be called "building", I
don't want this term to be overloaded with a completely different
meaning. I originally chose "build_*" because it aligned with the
FactoryBot definition (i.e. "build" is in memory whereas "create" is
in the database); since "build" means something else in the context
of this project, I'll go with "create" for now.