Introduce building of resources
I'm still not sure how this "pure data" (i.e. no filesystem) model of
resources will work, especially with things like incremental builds and
assets.

I like how the build step is just a data transform, and it gets written
to the filesystem at a different point in the process. However, this
separation could become troublesome and tedious. I might instead have
some kind of "source" on each resource from which the name, template,
etc. are derived.

How this would change resource building, I'm not sure. Maybe instead
of a pure data transform it becomes writing to some kind of "buffer",
which could be a file or something else.