Use string keys for metadata instead of symbols
I tend to think of symbols as keys in a hash which has a predefined
structure; think something parsed from a bunch of normalized JSON
where each object has the same keys. String keys, on the other hand,
are more indicative of a hash with "random" keys, or keys which have
no predefined structure or expected names.

I was initially drawn towards symbols because of specific, meaningful
keys like `:title` or `:layout`; now I realize that the majority of
metadata keys are arbitrary (as indicated by strings), and I'm just
*deriving* meaning from keys which don't inherently have meaning.

This is a little philosophical, but it's these kind of small
considerations that I think add up to create a coherent mental model.